Moved components, hooks, screens, and styles to 'src' folder
- Created a new folder named 'src' at the root of the project.
- Moved the 'components', 'hooks', 'screens', and 'styles' folders to the 'src' folder.
- Updated all file paths and import statements to reflect the new folder structure.
- Verified that the application builds and runs successfully after the folder relocation.
